requirements.txt
streamlit
numpy
scipy.optimize